refactor: finish getting devserver working again
